fix: fix some of the compilation errors and warnings

+++ src/MeoAssistant/InfrastPowerTask.cpp
@@ -9,7 +9,7 @@ bool asst::InfrastPowerTask::_run()
     // 发电站只能造这一个
     set_product("Drone");
 
-    for (; m_cur_facility_index != max_num_of_facilities(); ++m_cur_facility_index) {
+    for (; m_cur_facility_index != static_cast<int>(max_num_of_facilities()); ++m_cur_facility_index) {
         if (need_exit()) {
             return false;
         }
